2A01 Alarm
Introduce joinThread() which spawns a thread and immediately wait to join on said thread. The child threads still running when the alarm thread ends can either be cancelled (default) or simply interrupted. The result value is similar to a static join not including the alarm thread; when cancelled, threads have a special Timeout value, while interrupted threads just have whatever value they had when the alarm occurred. This allows a simple implementation of duration; first and join can implement max an min duration.