4F0I Either as a container (#47)
Either is a container allowing its contents to handle errors as well as
values. It can either handle both in the same path (the `error` property
of the fiber will be set if it is failing; the `value` property should
be accessed otherwise), or on two different paths (the first path will
only process values, as usual, while the second path will only process
errors).