4G0G Dynamic delay duration (#51)
Allow updating the current delay for a fiber from the scheduler. In
order to be able to access the fiber, also give a way to name fibers so
that they can be retrieved by their unique name using the `Fiber.byName`
map. Update the timer example to avoid the ugly loop condition check.