4G0D Ramp (#52)
A ramp is a kind of active delay with a delegate method that gets called
when the ramp starts and end, and after each update. This can be used
for smooth animations as the clock is based on requestAnimationFrame.
The delegate method `rampDidProgress` is called with a `p` parameter,
with guarantees that the first and last call are with values 0 and 1,
and all calls in between with values between 0 and 1. The duration can
be updated with `updateDelay` as well. The timer example is improved by
adding a smoother animation of the progress bar and simpler code.