4C08 Fiber rate (#57)
Add a rate property to the fiber that divides durations of delays and
ramps (a higher rate makes durations shorter; a lower rate makes
durations longer). Supported values are finite values greater than 0 (0,
∞ and negative values will be addressed later). One difficulty is
rescheduling current delays/ramps. Events and asynchronous function
calls are not affected. The turtle demo is updated to allow the user to
set the turtle speed while it is running.